requirements.txt
Flask==3.0.2
Pillow==10.3.0
numpy==1.26.4
Werkzeug==3.0.1